pyproject.toml
[tool.poetry]
name = "vcbackend"
version = "0.1.0"
description = ""
authors = ["Caio <caiaoviniciuscc@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.34.0"
fastapi = "^0.115.6"
sqlalchemy = "^2.0.36"
sqlmodel = "^0.0.22"
python-dotenv = "^1.0.1"
psycopg = {version = "^3.2.1", extras = ["binary"]}
boto3 = "^1.36.2"
python-multipart = "^0.0.20"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
